v6 Dev (#589)
Refactor to DDD

+++ app/Http/Controllers/AuthController.php
@@ -5,26 +5,26 @@ declare(strict_types=1);
 namespace Kami\Cocktail\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Kami\Cocktail\Models\User;
 use OpenApi\Attributes as OAT;
-use Illuminate\Http\JsonResponse;
 use Kami\Cocktail\OpenAPI as BAO;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Auth\Events\Verified;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Mail;
 use Kami\Cocktail\Mail\PasswordReset;
-use Kami\Cocktail\Mail\ConfirmAccount;
 use Kami\Cocktail\Mail\PasswordChanged;
 use Illuminate\Support\Facades\Password;
 use Kami\Cocktail\Http\Resources\TokenResource;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Kami\Cocktail\Http\Requests\RegisterRequest;
-use Kami\Cocktail\Http\Resources\ProfileResource;
+use Kami\Cocktail\Services\Auth\RegisterUserService;
+use Kami\Cocktail\OpenAPI\Schemas\RegisterRequest as SchemaRegisterRequest;
 
 class AuthController extends Controller
 {
-    #[OAT\Post(path: '/auth/login', tags: ['Authentication'], operationId: 'authenticate', summary: 'Authenticate user', description: 'Authenticate user and get auth token', requestBody: new OAT\RequestBody(
+    #[OAT\Post(path: '/auth/login', tags: ['Authentication'], operationId: 'authenticate', summary: 'Authenticate user', description: 'Authenticate user and get bearer token. Depending on server settings user might also require verified email.', requestBody: new OAT\RequestBody(
         required: true,
         content: [
             new OAT\JsonContent(ref: BAO\Schemas\LoginRequest::class),
@@ -73,55 +73,43 @@ class AuthController extends Controller
         return new TokenResource($token);
     }
 
-    #[OAT\Post(path: '/auth/logout', tags: ['Authentication'], operationId: 'logout', description: 'Logout currently authenticated user', summary: 'Logout')]
+    #[OAT\Post(path: '/auth/logout', tags: ['Authentication'], operationId: 'logout', description: 'Logout currently authenticated user. This will delete and disable bearer token.', summary: 'Logout')]
     #[OAT\Response(response: 204, description: 'Successful response')]
-    public function logout(Request $request): JsonResponse
+    public function logout(Request $request): Response
     {
         /** @var \Laravel\Sanctum\PersonalAccessToken */
         $currentAccessToken = $request->user()->currentAccessToken();
-        $currentAccessToken->delete();
+        if ($currentAccessToken) {
+            $currentAccessToken->delete();
+        }
 
-        return response()->json(status: 204);
+        return response()->noContent();
     }
 
-    #[OAT\Post(path: '/auth/register', tags: ['Authentication'], operationId: 'register', description: 'Register a new user', summary: 'Register', requestBody: new OAT\RequestBody(
+    #[OAT\Post(path: '/auth/register', tags: ['Authentication'], operationId: 'register', description: 'Register a new user account. If server has disabled registrations this will return 404 not found.', summary: 'Register', requestBody: new OAT\RequestBody(
         required: true,
         content: [
             new OAT\JsonContent(ref: BAO\Schemas\RegisterRequest::class),
         ]
     ), security: [])]
-    #[BAO\SuccessfulResponse(content: [
-        new BAO\WrapObjectWithData(ProfileResource::class),
+    #[BAO\SuccessfulResponse(response: 201, headers: [
+        new OAT\Header(header: 'Location', description: 'URL of the new resource', schema: new OAT\Schema(type: 'string')),
     ])]
     #[BAO\NotFoundResponse]
-    public function register(RegisterRequest $req): JsonResource
+    public function register(RegisterUserService $registerService, RegisterRequest $req): Response
     {
         if (config('bar-assistant.allow_registration') === false) {
             abort(404, 'Registrations are disabled.');
         }
 
-        $requireConfirmation = config('bar-assistant.mail_require_confirmation');
+        $user = $registerService->register(
+            SchemaRegisterRequest::fromIlluminateRequest($req)
+        );
 
-        $user = new User();
-        $user->name = $req->input('name');
-        $user->password = Hash::make($req->input('password'));
-        $user->email = $req->input('email');
-        if ($requireConfirmation === false) {
-            $user->email_verified_at = now();
-        }
-        $user->save();
-
-        if ($requireConfirmation === true) {
-            Mail::to($user)->queue(new ConfirmAccount($user->id, sha1((string) $user->email)));
-            Log::info('User registered and confirmation email sent', [
-                'email' => $user->email,
-            ]);
-        }
-
-        return new ProfileResource($user);
+        return new Response(status: 201, headers: ['Location' => route('profile.show', $user->id, false)]);
     }
 
-    #[OAT\Post(path: '/auth/forgot-password', tags: ['Authentication'], operationId: 'passwordForgot', description: 'Request a new password reset link', summary: 'Request password reset', requestBody: new OAT\RequestBody(
+    #[OAT\Post(path: '/auth/forgot-password', tags: ['Authentication'], operationId: 'passwordForgot', description: 'Send a new password reset link to users email address.', summary: 'Request password reset', requestBody: new OAT\RequestBody(
         required: true,
         content: [
             new OAT\JsonContent(type: 'object', properties: [
@@ -130,8 +118,7 @@ class AuthController extends Controller
         ]
     ), security: [])]
     #[OAT\Response(response: 204, description: 'Password reset link sent')]
-    #[OAT\Response(response: 400, description: 'Unable to send password reset link')]
-    public function passwordForgot(Request $request): JsonResponse
+    public function passwordForgot(Request $request): Response
     {
         $request->validate(['email' => 'required|email']);
 
@@ -145,13 +132,13 @@ class AuthController extends Controller
         });
 
         if ($status === Password::RESET_LINK_SENT) {
-            return response()->json(status: 204);
+            return response()->noContent();
         }
 
-        abort(400);
+        return response()->noContent();
     }
 
-    #[OAT\Post(path: '/auth/reset-password', tags: ['Authentication'], operationId: 'passwordReset', description: 'Reset user password', summary: 'Reset password', requestBody: new OAT\RequestBody(
+    #[OAT\Post(path: '/auth/reset-password', tags: ['Authentication'], operationId: 'passwordReset', description: 'Reset user password with data from password reset request.', summary: 'Reset password', requestBody: new OAT\RequestBody(
         required: true,
         content: [
             new OAT\JsonContent(type: 'object', properties: [
@@ -162,9 +149,9 @@ class AuthController extends Controller
             ]),
         ]
     ), security: [])]
-    #[OAT\Response(response: 204, description: 'Password succssfully reset')]
+    #[OAT\Response(response: 204, description: 'Password succesfully reset')]
     #[OAT\Response(response: 400, description: 'Unable to reset password')]
-    public function passwordReset(Request $request): JsonResponse
+    public function passwordReset(Request $request): Response
     {
         $request->validate([
             'token' => 'required',
@@ -187,20 +174,20 @@ class AuthController extends Controller
                 'email' => $request->post('email'),
             ]);
 
-            return response()->json(status: 204);
+            return response()->noContent();
         }
 
         abort(400, $status);
     }
 
-    #[OAT\Get(path: '/auth/verify/{id}/{hash}', tags: ['Authentication'], operationId: 'confirmAccount', description: 'Confirm user account, if applicable', summary: 'Confirm account', parameters: [
+    #[OAT\Get(path: '/auth/verify/{id}/{hash}', tags: ['Authentication'], operationId: 'confirmAccount', description: 'Verify user email. This will return 404 if email verifications is disabled.', summary: 'Confirm account', parameters: [
         new OAT\Parameter(name: 'id', in: 'path', required: true, description: 'Database id of a user', schema: new OAT\Schema(type: 'integer')),
         new OAT\Parameter(name: 'hash', in: 'path', required: true, description: 'Hash string sent to user email', schema: new OAT\Schema(type: 'string')),
     ], security: [])]
     #[OAT\Response(response: 204, description: 'Account confirmed')]
     #[BAO\NotAuthorizedResponse]
     #[BAO\NotFoundResponse]
-    public function confirmAccount(int $id, string $hash): JsonResponse
+    public function confirmAccount(int $id, string $hash): Response
     {
         if (config('bar-assistant.mail_require_confirmation') === false) {
             abort(404);
@@ -222,15 +209,6 @@ class AuthController extends Controller
             event(new Verified($user));
         }
 
-        return response()->json(status: 204);
-    }
-
-    public function passwordCheck(Request $request): JsonResponse
-    {
-        $status = Hash::check($request->input('password'), $request->user()->password);
-
-        return response()->json(['data' => [
-            'status' => $status,
-        ]]);
+        return response()->noContent();
     }
 }
